Color and performance
Added support for color! text and canvas only for now. Text is updating slowly for large images. and drop and drop isn't loading correctly when app first opens.

+++ src/app/stratton/gameoflife/renderers/WebGLRendererComponent.ts
@@ -138,7 +138,7 @@ export class WebGlRendererComponent implements Stratton.GameOfLife.IRenderer {
         });
     }
 
-    render(state: Int8Array, constraints: Stratton.GameOfLife.IConstraints) {
+    render(state: Int32Array, constraints: Stratton.GameOfLife.IConstraints) {
         if (!this.isInitialized) {
             return;
         }
